modules/translation/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Translation providers
google-cloud-translate==3.14.0
aiohttp==3.9.1

# Language detection
langdetect==1.0.9

# Celery and task queue
celery==5.3.6
redis==5.0.1

# Streamlit UI
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
